Rename common.cc et al to context_impl.cc et al.
