+++ source/common/config/BUILD
@@ -108,7 +108,7 @@ envoy_cc_library(
         "//source/common/common:backoff_lib",
         "//source/common/common:minimal_logger_lib",
         "//source/common/common:token_bucket_impl_lib",
-        "//source/common/grpc:common_lib",
+        "//source/common/grpc:context_lib",
         "//source/common/protobuf",
         "@envoy_api//envoy/api/v2:discovery_cc",
     ],
@@ -194,7 +194,7 @@ envoy_cc_library(
         "//include/envoy/event:dispatcher_interface",
         "//source/common/common:assert_lib",
         "//source/common/common:minimal_logger_lib",
-        "//source/common/grpc:common_lib",
+        "//source/common/grpc:context_lib",
         "//source/common/protobuf",
         "@envoy_api//envoy/api/v2:discovery_cc",
     ],
@@ -345,7 +345,7 @@ envoy_cc_library(
     srcs = ["type_to_endpoint.cc"],
     hdrs = ["type_to_endpoint.h"],
     deps = [
-        "//source/common/grpc:common_lib",
+        "//source/common/grpc:context_lib",
         "//source/common/protobuf",
         "@envoy_api//envoy/api/v2:cds_cc",
         "@envoy_api//envoy/api/v2:eds_cc",
@@ -371,7 +371,7 @@ envoy_cc_library(
         "//source/common/common:assert_lib",
         "//source/common/common:hash_lib",
         "//source/common/common:hex_lib",
-        "//source/common/grpc:common_lib",
+        "//source/common/grpc:context_lib",
         "//source/common/json:config_schemas_lib",
         "//source/common/protobuf",
         "//source/common/protobuf:utility_lib",
