+++ source/common/grpc/BUILD
@@ -15,7 +15,7 @@ envoy_cc_library(
     hdrs = ["typed_async_client.h"],
     deps = [
         ":codec_lib",
-        ":common_lib",
+        ":context_lib",
         "//include/envoy/grpc:async_client_interface",
         "//source/common/buffer:zero_copy_input_stream_lib",
         "//source/common/http:async_client_lib",
@@ -28,7 +28,7 @@ envoy_cc_library(
     hdrs = ["async_client_impl.h"],
     deps = [
         ":codec_lib",
-        ":common_lib",
+        ":context_lib",
         ":typed_async_client_lib",
         "//include/envoy/grpc:async_client_interface",
         "//source/common/buffer:zero_copy_input_stream_lib",
@@ -42,7 +42,7 @@ envoy_cc_library(
     hdrs = ["async_client_manager_impl.h"],
     deps = [
         ":async_client_lib",
-        ":common_lib",
+        ":context_lib",
         "//include/envoy/grpc:async_client_manager_interface",
         "//include/envoy/singleton:manager_interface",
         "//include/envoy/thread_local:thread_local_interface",
@@ -73,9 +73,9 @@ envoy_cc_library(
 )
 
 envoy_cc_library(
-    name = "common_lib",
-    srcs = ["common.cc"],
-    hdrs = ["common.h"],
+    name = "context_lib",
+    srcs = ["context_impl.cc"],
+    hdrs = ["context_impl.h"],
     external_deps = ["abseil_optional"],
     deps = [
         "//include/envoy/grpc:context_interface",
@@ -129,7 +129,7 @@ envoy_cc_library(
         "grpc",
     ],
     deps = [
-        ":common_lib",
+        ":context_lib",
         ":google_grpc_creds_lib",
         ":google_grpc_utils_lib",
         ":typed_async_client_lib",
