+++ source/common/grpc/google_async_client_impl.cc
@@ -5,7 +5,7 @@
 #include "common/common/empty_string.h"
 #include "common/common/lock_guard.h"
 #include "common/config/datasource.h"
-#include "common/grpc/common.h"
+#include "common/grpc/context_impl.h"
 #include "common/grpc/google_grpc_creds_impl.h"
 #include "common/grpc/google_grpc_utils.h"
 #include "common/tracing/http_tracer_impl.h"
