+++ source/common/router/retry_state_impl.cc
@@ -7,7 +7,7 @@
 
 #include "common/common/assert.h"
 #include "common/common/utility.h"
-#include "common/grpc/common.h"
+#include "common/grpc/context_impl.h"
 #include "common/http/codes.h"
 #include "common/http/headers.h"
 #include "common/http/utility.h"
