+++ source/extensions/filters/http/grpc_http1_bridge/http1_bridge_filter.cc
@@ -8,7 +8,7 @@
 
 #include "common/common/enum_to_int.h"
 #include "common/common/utility.h"
-#include "common/grpc/common.h"
+#include "common/grpc/context_impl.h"
 #include "common/http/headers.h"
 #include "common/http/http1/codec_impl.h"
