+++ source/extensions/filters/http/grpc_http1_reverse_bridge/BUILD
@@ -16,7 +16,7 @@ envoy_cc_library(
         "//include/envoy/http:filter_interface",
         "//source/common/common:enum_to_int",
         "//source/common/grpc:codec_lib",
-        "//source/common/grpc:common_lib",
+        "//source/common/grpc:context_lib",
         "//source/common/grpc:status_lib",
         "//source/common/http:header_map_lib",
         "//source/extensions/filters/http:well_known_names",
