+++ source/extensions/filters/http/grpc_json_transcoder/BUILD
@@ -26,7 +26,7 @@ envoy_cc_library(
         ":transcoder_input_stream_lib",
         "//include/envoy/http:filter_interface",
         "//source/common/grpc:codec_lib",
-        "//source/common/grpc:common_lib",
+        "//source/common/grpc:context_lib",
         "//source/common/http:headers_lib",
         "//source/common/protobuf",
         "@envoy_api//envoy/config/filter/http/transcoder/v2:transcoder_cc",
