+++ source/extensions/filters/http/grpc_web/grpc_web_filter.cc
@@ -6,7 +6,7 @@
 #include "common/common/base64.h"
 #include "common/common/empty_string.h"
 #include "common/common/utility.h"
-#include "common/grpc/common.h"
+#include "common/grpc/context_impl.h"
 #include "common/http/headers.h"
 #include "common/http/utility.h"
