+++ source/extensions/tracers/lightstep/lightstep_tracer_impl.cc
@@ -9,7 +9,7 @@
 #include "common/common/base64.h"
 #include "common/common/fmt.h"
 #include "common/config/utility.h"
-#include "common/grpc/common.h"
+#include "common/grpc/context_impl.h"
 #include "common/http/message_impl.h"
 #include "common/tracing/http_tracer_impl.h"
