+++ source/server/config_validation/server.h
@@ -10,7 +10,7 @@
 
 #include "common/access_log/access_log_manager_impl.h"
 #include "common/common/assert.h"
-#include "common/grpc/common.h"
+#include "common/grpc/context_impl.h"
 #include "common/protobuf/message_validator_impl.h"
 #include "common/router/rds_impl.h"
 #include "common/runtime/runtime_impl.h"
