+++ test/common/grpc/context_impl_test.cc
@@ -1,6 +1,6 @@
 #include <arpa/inet.h>
 
-#include "common/grpc/common.h"
+#include "common/grpc/context_impl.h"
 #include "common/http/headers.h"
 #include "common/http/message_impl.h"
 #include "common/http/utility.h"
