+++ test/extensions/access_loggers/http_grpc/BUILD
@@ -43,7 +43,7 @@ envoy_extension_cc_test(
     deps = [
         "//source/common/buffer:zero_copy_input_stream_lib",
         "//source/common/grpc:codec_lib",
-        "//source/common/grpc:common_lib",
+        "//source/common/grpc:context_lib",
         "//source/extensions/access_loggers/http_grpc:config",
         "//test/common/grpc:grpc_client_integration_lib",
         "//test/integration:http_integration_lib",
