+++ test/extensions/access_loggers/http_grpc/grpc_access_log_integration_test.cc
@@ -4,7 +4,7 @@
 #include "common/buffer/zero_copy_input_stream_impl.h"
 #include "common/common/version.h"
 #include "common/grpc/codec.h"
-#include "common/grpc/common.h"
+#include "common/grpc/context_impl.h"
 
 #include "test/common/grpc/grpc_client_integration.h"
 #include "test/integration/http_integration.h"
