+++ test/extensions/filters/http/grpc_json_transcoder/grpc_json_transcoder_integration_test.cc
@@ -1,5 +1,5 @@
 #include "common/grpc/codec.h"
-#include "common/grpc/common.h"
+#include "common/grpc/context_impl.h"
 #include "common/http/message_impl.h"
 #include "common/protobuf/protobuf.h"
