+++ test/extensions/filters/http/grpc_json_transcoder/json_transcoder_filter_test.cc
@@ -4,7 +4,7 @@
 #include "common/buffer/buffer_impl.h"
 #include "common/config/filter_json.h"
 #include "common/grpc/codec.h"
-#include "common/grpc/common.h"
+#include "common/grpc/context_impl.h"
 #include "common/http/header_map_impl.h"
 #include "common/protobuf/protobuf.h"
