+++ test/extensions/filters/http/grpc_web/grpc_web_filter_test.cc
@@ -3,7 +3,7 @@
 #include "common/buffer/buffer_impl.h"
 #include "common/common/base64.h"
 #include "common/common/utility.h"
-#include "common/grpc/common.h"
+#include "common/grpc/context_impl.h"
 #include "common/http/codes.h"
 #include "common/http/header_map_impl.h"
 #include "common/http/headers.h"
