+++ test/extensions/stats_sinks/metrics_service/metrics_service_integration_test.cc
@@ -3,7 +3,7 @@
 
 #include "common/common/version.h"
 #include "common/grpc/codec.h"
-#include "common/grpc/common.h"
+#include "common/grpc/context_impl.h"
 #include "common/stats/histogram_impl.h"
 
 #include "test/common/grpc/grpc_client_integration.h"
