+++ test/extensions/tracers/lightstep/lightstep_tracer_impl_test.cc
@@ -4,7 +4,7 @@
 #include <string>
 
 #include "common/common/base64.h"
-#include "common/grpc/common.h"
+#include "common/grpc/context_impl.h"
 #include "common/http/header_map_impl.h"
 #include "common/http/headers.h"
 #include "common/http/message_impl.h"
