+++ test/integration/BUILD
@@ -390,7 +390,7 @@ envoy_cc_test_library(
         "//source/common/common:minimal_logger_lib",
         "//source/common/event:dispatcher_lib",
         "//source/common/grpc:codec_lib",
-        "//source/common/grpc:common_lib",
+        "//source/common/grpc:context_lib",
         "//source/common/http:codec_client_lib",
         "//source/common/http:header_map_lib",
         "//source/common/http:headers_lib",
@@ -590,7 +590,7 @@ envoy_cc_test(
         ":http_integration_lib",
         "//source/common/buffer:zero_copy_input_stream_lib",
         "//source/common/grpc:codec_lib",
-        "//source/common/grpc:common_lib",
+        "//source/common/grpc:context_lib",
         "//source/extensions/filters/http/ratelimit:config",
         "//test/common/grpc:grpc_client_integration_lib",
         "@envoy_api//envoy/service/ratelimit/v2:rls_cc",
