+++ test/integration/ratelimit_integration_test.cc
@@ -2,7 +2,7 @@
 
 #include "common/buffer/zero_copy_input_stream_impl.h"
 #include "common/grpc/codec.h"
-#include "common/grpc/common.h"
+#include "common/grpc/context_impl.h"
 
 #include "extensions/filters/http/ratelimit/config.h"
